+++ src/entry.rs
@@ -3,6 +3,8 @@ use std::option::Option;
 use std::path::Path;
 use std::path::PathBuf;
 
+use crate::SyncOptions;
+
 #[derive(Debug, Clone)]
 pub struct Entry {
     description: String,
@@ -10,8 +12,8 @@ pub struct Entry {
     metadata: Option<fs::Metadata>,
     exists: bool,
     is_link: Option<bool>,
-    offset: Option<usize>,
-    length: Option<usize>,
+    chunk_offset: Option<usize>,
+    chunk_length: Option<usize>,
 }
 
 impl Entry {
@@ -32,15 +34,15 @@ impl Entry {
             path: entry_path.to_path_buf(),
             exists: entry_path.exists(),
             is_link,
-            offset: None,
-            length: None,
+            chunk_offset: None,
+            chunk_length: None,
         }
     }
 
     pub fn to_chunk(&self, offset: usize, length: usize) -> Entry {
         let mut entry = self.clone();
-        entry.offset = Some(offset);
-        entry.length = Some(length);
+        entry.chunk_offset = Some(offset);
+        entry.chunk_length = Some(length);
         entry
     }
 
@@ -62,16 +64,30 @@ impl Entry {
         self.is_link
     }
 
-    pub fn get_offset(&self) -> Option<usize> {
-        self.offset
+    pub fn is_file(&self) -> Option<bool> {
+        if self.metadata.is_none() {
+            return None;
+        }
+        Some(self.metadata().unwrap().is_file())
     }
 
-    pub fn get_length(&self) -> Option<usize> {
-        self.length
+    pub fn length(&self) -> Option<usize> {
+        if self.metadata.is_none() {
+            return None;
+        }
+        Some(self.metadata().unwrap().len() as usize)
+    }
+
+    pub fn chunk_offset(&self) -> Option<usize> {
+        self.chunk_offset
+    }
+
+    pub fn chunk_length(&self) -> Option<usize> {
+        self.chunk_length
     }
 
     pub fn is_chunk(&self) -> bool {
-        self.offset.is_some() && self.length.is_some()
+        self.chunk_offset.is_some() && self.chunk_length.is_some()
     }
 
     // pub fn is_first_chunk(&self) -> bool {
@@ -79,7 +95,7 @@ impl Entry {
     // }
 
     pub fn is_last_chunk(&self) -> bool {
-        self.offset == Some(self.metadata.as_ref().unwrap().len() as usize - self.length.unwrap())
+        self.chunk_offset == Some(self.length().unwrap() - self.chunk_length.unwrap())
     }
 }
 
@@ -109,4 +125,10 @@ mod tests {
         assert!(is_link.is_some());
         assert!(!is_link.unwrap());
     }
+}
+
+pub struct CopyEntry {
+    pub src: Entry,
+    pub dest: Entry,
+    pub opts: SyncOptions,
 }